# OpenAPI.yaml
openapi: 3.0.0
info:
  title: store
  version: '1.1'
  description: Store API for customers, orders, and products.
  contact:
    name: SecuritEase Dev
    url: 'https://www.securitease.com'
    email: internal@securitease.com
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.htm'
servers:
  - url: 'http://localhost:8080'
    description: dev
paths:
  /health:
    get:
      summary: Application health check
      tags: [store]
      operationId: getHealth
      responses:
        '200':
          description: Application is running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /order:
    get:
      summary: Get orders
      tags: [store]
      operationId: getOrders
      description: Returns a paginated list of orders with their customer summary and product summaries.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderPage'
    post:
      summary: Create order
      tags: [store]
      operationId: createOrder
      description: Creates an order for an existing customer and one or more existing products.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer or product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /order/{id}:
    get:
      summary: Get order by ID
      tags: [store]
      operationId: getOrderById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customer:
    get:
      summary: Get customers
      tags: [store]
      operationId: getCustomers
      description: Returns a paginated list of customers, or filters by case-insensitive substring using the q parameter.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Case-insensitive substring to match against a customer name.
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPage'
    post:
      summary: Create customer
      tags: [store]
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customer/{id}:
    get:
      summary: Get customer by ID
      tags: [store]
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid customer ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products:
    get:
      summary: Get products
      tags: [store]
      operationId: getProducts
      description: Returns a paginated list of products and the order IDs containing each product.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'
    post:
      summary: Create product
      tags: [store]
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /products/{id}:
    get:
      summary: Get product by ID
      tags: [store]
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Zero-based page number.
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Number of records per page.
  schemas:
    CreateCustomerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: Jane Doe
    CreateProductRequest:
      type: object
      required: [description]
      properties:
        description:
          type: string
          example: Mechanical Keyboard
    CreateOrderRequest:
      type: object
      required: [description, customerId, productIds]
      properties:
        description:
          type: string
          example: Home office setup
        customerId:
          type: integer
          format: int64
          example: 1
        productIds:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
          example: [1, 2, 4]
    Order:
      title: Order
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        customer:
          $ref: '#/components/schemas/OrderCustomer'
        products:
          type: array
          items:
            $ref: '#/components/schemas/OrderProduct'
    OrderCustomer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    OrderProduct:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
    Customer:
      title: Customer
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        orders:
          type: array
          items:
            $ref: '#/components/schemas/CustomerOrder'
    CustomerOrder:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
    Product:
      title: Product
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        orderIds:
          type: array
          items:
            type: integer
            format: int64
    OrderPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Order'
    CustomerPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Customer'
    ProductPage:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/Product'
    PageMetadata:
      type: object
      properties:
        page:
          type: integer
          example: 0
        size:
          type: integer
          example: 50
        totalElements:
          type: integer
          format: int64
          example: 125
        totalPages:
          type: integer
          example: 3
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Validation failed
        path:
          type: string
          example: /products
        details:
          nullable: true
          description: Optional structured error details.
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: UP
        service:
          type: string
          example: store
        checkedAt:
          type: string
          format: date-time
tags:
  - name: store
    description: Store operations